Remove old payment gateway task components
