+++ client/payments/payment-recommendations.tsx
@@ -32,7 +32,7 @@ import './payment-recommendations.scss';
 import { getCountryCode } from '../dashboard/utils';
 import { getAdminLink } from '../wc-admin-settings';
 import { createNoticesFromResponse } from '../lib/notices';
-import { isWCPaySupported } from '../task-list/tasks/payments/methods/wcpay';
+import { isWCPaySupported } from '../task-list/tasks/payments/PaymentGatewaySuggestions/components/WCPay';
 
 const SEE_MORE_LINK =
 	'https://woocommerce.com/product-category/woocommerce-extensions/payment-gateways/?utm_source=payments_recommendations';
