+++ client/profile-wizard/steps/business-details/flows/selective-bundle/selective-extensions-bundle/index.js
@@ -20,7 +20,7 @@ import { AppIllustration } from '../app-illustration';
 import './style.scss';
 import { setAllPropsToValue } from '~/lib/collections';
 import { getCountryCode } from '~/dashboard/utils';
-import { isWCPaySupported } from '~/task-list/tasks/payments/methods/wcpay';
+import { isWCPaySupported } from '~/task-list/tasks/payments/PaymentGatewaySuggestions/components/WCPay';
 
 const generatePluginDescriptionWithLink = (
 	description,
