+++ client/task-list/tasks/payments/PaymentGatewaySuggestions/components/WCPay/index.js
@@ -1,0 +1,3 @@
+export * from './utils';
+export { Suggestion as WCPaySuggestion } from './Suggestion';
+export { UsageModal as WCPayUsageModal } from './UsageModal';
