+++ client/task-list/tasks/payments/PaymentGatewaySuggestions/components/WCPay/utils.js
@@ -48,3 +48,24 @@ export function installActivateAndConnectWcpay(
 			reject();
 		} );
 }
+
+export function isWCPaySupported( countryCode ) {
+	const supportedCountries = [ 'US', 'PR' ];
+	if (
+		window.wcAdminFeatures &&
+		window.wcAdminFeatures[ 'wcpay/support-international-countries' ]
+	) {
+		supportedCountries.push(
+			'AU',
+			'CA',
+			'DE',
+			'ES',
+			'FR',
+			'GB',
+			'IE',
+			'IT',
+			'NZ'
+		);
+	}
+	return supportedCountries.includes( countryCode );
+}
