--- client/task-list/tasks/payments/components/PaymentMethodList/PaymentMethodList.scss
@@ -1,81 +1,0 @@
-.woocommerce-task-payment {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	position: relative;
-	overflow: hidden;
-
-	.components-card__media {
-		width: 170px;
-		flex-shrink: 0;
-
-		img,
-		svg {
-			margin: auto;
-			max-width: 100px;
-			display: block;
-		}
-	}
-
-	> .components-form-toggle {
-		min-width: 52px;
-	}
-
-	.woocommerce-task-payment__title {
-		display: flex;
-		align-items: center;
-		font-size: 16px;
-		font-weight: 500;
-		color: $studio-gray-80;
-		margin-top: 0;
-		margin-bottom: $gap-smaller;
-	}
-
-	.woocommerce-task-payment__content {
-		font-size: 14px;
-		color: $studio-gray-60;
-		margin: 0;
-		margin-right: $gap-larger;
-
-		.text-style-strong {
-			font-weight: bold;
-		}
-
-		p {
-			font-size: 12px;
-		}
-	}
-	.woocommerce-task-payment__description {
-		flex: 1;
-	}
-
-	@include breakpoint( '<600px' ) {
-		flex-wrap: wrap;
-
-		.woocommerce-task-payment__content {
-			margin: 0;
-		}
-
-		.components-card__media {
-			> svg,
-			> img {
-				margin: 0 0 0 $gap-large;
-			}
-
-			order: 1;
-			flex-basis: 50%;
-		}
-
-		.woocommerce-task-payment__description {
-			order: 3;
-			padding: $gap-large 0 0 $gap-large;
-		}
-
-		.woocommerce-task-payment__footer {
-			flex-basis: 50%;
-			align-self: flex-start;
-			order: 2;
-			text-align: right;
-		}
-	}
-}
