--- client/task-list/tasks/payments/components/PaymentMethodList/index.js
@@ -1,1 +1,0 @@
-export { PaymentMethodList } from './PaymentMethodList';
