--- client/task-list/tasks/payments/images/bacs.js
@@ -1,74 +1,0 @@
-export default () => (
-	<svg
-		width="96"
-		height="32"
-		viewBox="0 0 96 32"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<rect width="32" height="32" rx="16" fill="#8E9196" />
-		<mask
-			id="bacs0"
-			mask-type="alpha"
-			maskUnits="userSpaceOnUse"
-			x="8"
-			y="8"
-			width="16"
-			height="16"
-		>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M8.875 12.25L16 8.5L23.125 12.25V13.75H8.875V12.25ZM16 10.195L19.9075 12.25H12.0925L16 10.195ZM10.75 15.25H12.25V20.5H10.75V15.25ZM15.25 20.5V15.25H16.75V20.5H15.25ZM23.125 23.5V22H8.875V23.5H23.125ZM19.75 15.25H21.25V20.5H19.75V15.25Z"
-				fill="white"
-			/>
-		</mask>
-		<g mask="url(#bacs0)">
-			<rect x="7" y="7" width="18" height="18" fill="white" />
-		</g>
-		<mask
-			id="bacs1"
-			mask-type="alpha"
-			maskUnits="userSpaceOnUse"
-			x="39"
-			y="10"
-			width="18"
-			height="12"
-		>
-			<path
-				d="M39 17L53.17 17L49.59 20.59L51 22L57 16L51 10L49.59 11.41L53.17 15L39 15L39 17Z"
-				fill="white"
-			/>
-		</mask>
-		<g mask="url(#bacs1)">
-			<rect
-				x="60"
-				y="28"
-				width="24"
-				height="24"
-				transform="rotate(-180 60 28)"
-				fill="#8E9196"
-			/>
-		</g>
-		<rect x="64" width="32" height="32" rx="16" fill="#8E9196" />
-		<mask
-			id="bacs2"
-			mask-type="alpha"
-			maskUnits="userSpaceOnUse"
-			x="72"
-			y="8"
-			width="16"
-			height="16"
-		>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M72.875 12.25L80 8.5L87.125 12.25V13.75H72.875V12.25ZM80 10.195L83.9075 12.25H76.0925L80 10.195ZM74.75 15.25H76.25V20.5H74.75V15.25ZM79.25 20.5V15.25H80.75V20.5H79.25ZM87.125 23.5V22H72.875V23.5H87.125ZM83.75 15.25H85.25V20.5H83.75V15.25Z"
-				fill="white"
-			/>
-		</mask>
-		<g mask="url(#bacs2)">
-			<rect x="71" y="7" width="18" height="18" fill="white" />
-		</g>
-	</svg>
-);
