+++ client/task-list/tasks/payments/index.js
@@ -1,15 +1,10 @@
 /**
  * Internal dependencies
  */
-import { LocalPayments } from './LocalPayments';
 import { PaymentGatewaySuggestions } from './PaymentGatewaySuggestions';
 
 export const Payments = ( { query } ) => {
-	if ( window.wcAdminFeatures[ 'payment-gateway-suggestions' ] ) {
-		return <PaymentGatewaySuggestions query={ query } />;
-	}
-
-	return <LocalPayments query={ query } />;
+	return <PaymentGatewaySuggestions query={ query } />;
 };
 
 export default Payments;
