--- client/task-list/tasks/payments/methods/wcpay/index.js
@@ -1,10 +1,0 @@
-/**
- * Internal dependencies
- */
-import WCPayUsageModal from './wcpay-usage-modal';
-import WCPay from './wcpay';
-
-export * from './is-supported';
-export * from './install-activate-and-connect';
-
-export { WCPay, WCPayUsageModal };
