--- client/task-list/tasks/payments/methods/wcpay/is-supported.js
@@ -1,20 +1,0 @@
-export function isWCPaySupported( countryCode ) {
-	const supportedCountries = [ 'US', 'PR' ];
-	if (
-		window.wcAdminFeatures &&
-		window.wcAdminFeatures[ 'wcpay/support-international-countries' ]
-	) {
-		supportedCountries.push(
-			'AU',
-			'CA',
-			'DE',
-			'ES',
-			'FR',
-			'GB',
-			'IE',
-			'IT',
-			'NZ'
-		);
-	}
-	return supportedCountries.includes( countryCode );
-}
